Add files via upload
Hardwood colour trends seem to go in and out of style just like clothing trends do. Though the darks and lights are still the go-tos, it’s all about the different shades. Are your floors on-trend, or is it time for an update? Refinishing your floors can be an easy task if you have the right company working for you. Check out the top colours for 2020 and see if you need to update your floors heading into the New Year. Dark colours, to most people, are the most stylish. Several dark tones are very popular and look nice in any home. Not to mention they hide more grim and scuffs.